+++ NEWS.md
@@ -12,6 +12,8 @@
 
 * `link_gh_issue()` and `markup_href()` are better.
 
+* `proj_file()` is less noisy and is a shortcut for `file_outline()` with `proj`.
+
 # reuseme 0.0.1
 
 * in `file/proj/dir_outline()` `regex_outline` is now `pattern`